+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 ## Last Changes
 
+- [#2](https://github.com/LaxarJS/generator-laxarjs/issues/2): created one configuration file for bower.json defaults for the scaffolds